Relicense Sparkle to AGPL
The Section 13 provision of AGPL closes a loophole that would have allowed anyone to take the software and run it on a server in a closed fashion without needing to provide source as they were not distributing it. AGPL has a specific provision that prevents this.

More detail here: https://opensource.stackexchange.com/questions/5041/of-the-differences-between-the-gplv3-and-the-agplv3-texts-what-to-make-of-them