Fix crash in events
TypeScript did not catch that Record<> can return undefined if the key is not present. This appears to be a deficiency in the Record<> type. More info here: https://fnune.com/typescript/2019/01/30/typescript-series-1-record-is-usually-not-the-best-choice/

This fix wraps the cast in a Partial<> to force handling of the undefined case. Fortunately, no other areas were affeted by this specific bug.