wrap "Enter" into <a> to be able to open new tab/window
Such UX (ability to open a room in a new tab) is desperately needed, in particular
for poster rooms etc, since opening a room can take considerable time of a blue window
of sparkle.

I am not a web developer so do not know if this is a best way (most likely not),
and I have not added any guards etc.  May be ideas on
https://stackoverflow.com/a/2906586/1265472
could provide a better solution (e.g. making <button> into <a> with button styling etc)