ENH: show poster number: prior title in the poster preview and view
Poster number is important information since people refer to their posters by
numbers for easy finding/association/check if in the right place.  But it was
nowhere shown.

I guess ideally it should not be melded into the same <element> as title, but
this is the fastest and easiest way since title is in <p> and we do not want to
break it all across multiple lines etc.  Specific recommendations or better
tune up commits pushed directly would be most welcome to expedite this
PR.

Search for number yet to be supported (will see if could do in #1443) and is
orthogonal IMHO to having them displayed.