RF calendar related functions
Mostly stylistic changes.

I decided to drop the webcal protocol improvement.
With a bit of search around, there's no obvious way to do so with blob.
Most solution involves directly parsing the calendar info to the API of
different services. And there's other libraries for achiving this.

It's possible to set protocol with navigator.registerProtocolHandler,
but that introduce it's own issue that I cannot systematically
reporduce.