RF: do not disable "See all" for user list
From looking at the code, if disableSeeAll is True, all users
are shown.  At OHBM we will have > 1000 and that leads to
a very long list!  I think it would be useful for the event to
just collapse users by default and allow to expand if someone
really desires it

Closes #1579