npm uninstall jquery
Note that even though we get the following peerDependency warning, it's not relevant to us, as we are only using 'bootstrap' for it's SCSS, and react-bootstrap (which we use for the rest of the JS-related functionality) doesn't require jQuery

npm WARN bootstrap@4.5.0 requires a peer of jquery@1.9.1 - 3 but none is installed. You must install peer dependencies yourself.