Fixed UserReactions performance issue.
When rendered in large numbers, specifically inside a the Audience component, UserReactions has significant CPU cost.
This was caused because a custom css variable was created for every seated attendant using `useCss`.

This change fixes the problem by preventing the rendering of unnecessary DOM elements and removing the custom CSS variable.